test: measure no-boolean-toggle scaling by interleaved median
The min-of-N estimator is biased toward the shorter input under
intermittent runner contention: a ~25ms sample can land in a quiet
scheduling window that a ~125ms sample never fits inside, so
min(large)/min(small) inflates past the 18x bound even though the rule
is linear (~12us per setter from 1k to 20k setters). That is the exact
shape of the macOS CI failure (2k at idle speed, 10k 4x slower per
setter across all samples).

Keep the interleaved sampling, take the median of each size instead of
the minimum, and move the guard into its own .performance.test.ts like
the other scaling tests so it runs in a fresh worker. Disabling the
await-proof cache still trips the bound at ~25x.
